requirements.txt
boto3==1.37.38
python-dotenv==1.0.1
requests==2.32.3
